# docs/api/openapi.yml
openapi: 3.0.3
info:
  title: Algorithmic Trading System API
  description: |
    Enterprise-grade algorithmic trading platform API providing comprehensive trading, portfolio management, 
    risk management, and market data services. Built with microservices architecture and event-driven design.
    
    ## Key Features
    - **Multi-Asset Trading**: Stocks, ETFs, Options, Futures, Forex, Crypto
    - **Real-time Market Data**: Live feeds with fallback mechanisms
    - **Advanced Risk Management**: Real-time monitoring, VaR, stress testing
    - **AI-Powered Analytics**: Strategy development and optimization
    - **High-Performance Execution**: Microsecond-level latency
    
    ## Authentication
    All API endpoints require authentication via JWT tokens or API keys.
    
    ## Rate Limiting
    - Standard endpoints: 1000 requests/minute
    - Market data endpoints: 10000 requests/minute
    - Trading endpoints: 500 requests/minute
    
    ## Error Handling
    The API uses standard HTTP status codes and returns detailed error information in JSON format.
  version: 1.0.0
  contact:
    name: Trading System API Support
    email: api-support@trading-system.com
    url: https://docs.trading-system.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: https://api.trading-system.com/v1
    description: Production server
  - url: https://staging-api.trading-system.com/v1
    description: Staging server
  - url: http://localhost:8000/v1
...
security:
  - BearerAuth: []
  - ApiKeyAuth: []
paths:
...
  /auth/login:
    post:
      tags: [Authentication]
      summary: User login
      description: Authenticate user and return JWT token
      operationId: login
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: Login successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '401':
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /auth/refresh:
    post:
      tags: [Authentication]
      summary: Refresh JWT token
      description: Refresh an expired JWT token
      operationId: refreshToken
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RefreshTokenRequest'
      responses:
        '200':
          description: Token refreshed successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'

  # Market Data
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
  schemas:
...
    LoginRequest:
      type: object
      required: [username, password]
      properties:
        username:
          type: string
        password:
          type: string
          format: password
    RefreshTokenRequest:
      type: object
      required: [refresh_token]
      properties:
        refresh_token:
          type: string
    AuthResponse:
      type: object
      properties:
        access_token:
          type: string
        refresh_token:
          type: string
        token_type:
          type: string
          default: bearer
        expires_in:
          type: integer
        user:
          $ref: '#/components/schemas/User'
    User:
      type: object
      properties:
        id:
          type: string
          format: uuid
        username:
          type: string
        email:
          type: string
          format: email
        role:
          type: string
          enum: [admin, trader, analyst, viewer]
        created_at:
          type: string
          format: date-time

    # Market Data Schemas
...
    Error:
      type: object
      properties:
        error:
          type: string
        message:
          type: string
        code:
          type: integer
        timestamp:
          type: string
          format: date-time
        request_id:
          type: string
          format: uuid
...